Move some web stuff to http_stuff.go
